👌 Move app sync code to pkg/commands
